pyproject.toml
[project]
name = "claire-kg"
version = "0.1.0"
description = "CLAIRE Knowledge Graph - Cybersecurity Data Ingestion and Analysis"
authors = [
    {name = "Joel", email = "joel@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "GPL-3.0-or-later"}
dependencies = [
    "neo4j>=5.0.0",
    # "click>=8.0.0",  # not imported in src; typer brings it in as dependency
    # "pydantic>=2.0.0",  # not imported in src
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "httpx>=0.24.0",
    # "numpy>=1.24.0",  # not imported in src
    # "pandas>=2.0.0",  # not imported in src
    "tqdm>=4.65.0",
    # "openpyxl>=3.1.5",  # not imported in src
    "scikit-learn>=1.3.2",
    "openai>=2.6.1",
    "deepeval>=3.6.9",
    "tiktoken>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "-v"
]
markers = [
    "regression: Regression tests for existing functionality",
    "unit: Unit tests for new components",
    "integration: Integration tests for complete workflows",
    "slow: Slow running tests",
    "llm: Tests that require LLM API calls"
]

[dependency-groups]
dev = [
    "playwright>=1.56.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.3",
]
